chore: TSR-types: move all device-type definitions into the same folder
