+++ packages/timeline-state-resolver-types/src/integrations/atem.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from '.'
+import { DeviceType } from '..'
 
 export enum TimelineContentTypeAtem { //  Atem-state
 	ME = 'me',
