+++ packages/timeline-state-resolver-types/src/integrations/casparcg.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from '.'
+import { DeviceType } from '..'
 
 export enum TimelineContentTypeCasparCg { //  CasparCG-state
 	MEDIA = 'media',
