+++ packages/timeline-state-resolver-types/src/integrations/multiOsc.ts
