+++ packages/timeline-state-resolver-types/src/integrations/panasonicPTZ.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from '.'
+import { DeviceType } from '..'
 
 export enum TimelineContentTypePanasonicPtz {
 	PRESET = 'presetMem',
