+++ packages/timeline-state-resolver-types/src/integrations/pharos.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from '.'
+import { DeviceType } from '..'
 
 export enum TimelineContentTypePharos {
 	SCENE = 'scene',
