+++ packages/timeline-state-resolver-types/src/integrations/singularLive.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from '.'
+import { DeviceType } from '..'
 
 export interface SingularLiveContent {
 	type: TimelineContentTypeSingularLive
