+++ packages/timeline-state-resolver-types/src/integrations/sisyfos.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from '.'
+import { DeviceType } from '..'
 
 export enum TimelineContentTypeSisyfos {
 	CHANNEL = 'channel',
