+++ packages/timeline-state-resolver-types/src/integrations/sofieChef.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from '.'
+import { DeviceType } from '..'
 
 export enum TimelineContentTypeSofieChef {
 	URL = 'url',
