+++ packages/timeline-state-resolver-types/src/integrations/vmix.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from '.'
+import { DeviceType } from '..'
 
 export enum VMixCommand {
 	PREVIEW_INPUT = 'PREVIEW_INPUT',
